Introducing `BmModel` for the new ParticipantAgent
